requirements.txt
fastapi[standard]
numpy==2.2.2
pillow==11.1.0
python-multipart==0.0.20
scipy==1.15.1
xgboost==2.1.3
fastapi[standard]